Migrate list-filters
